+++ test/utils/test_dropout.py
@@ -1,6 +1,6 @@
 import torch
 
-from torch_geometric.utils import dropout_adj, dropout_node
+from torch_geometric.utils import dropout_adj, dropout_edge, dropout_node
 
 
 def test_dropout_adj():
@@ -41,3 +41,21 @@ def test_dropout_node():
     assert out[0].tolist() == [[2, 3], [3, 2]]
     assert out[1].tolist() == [False, False, False, False, True, True]
     assert out[2].tolist() == [True, False, True, True]
+
+
+def test_dropout_edge():
+    edge_index = torch.tensor([[0, 1, 1, 2, 2, 3], [1, 0, 2, 1, 3, 2]])
+
+    out = dropout_edge(edge_index, training=False)
+    assert edge_index.tolist() == out[0].tolist()
+    assert out[1].tolist() == [True, True, True, True, True, True]
+
+    torch.manual_seed(5)
+    out = dropout_edge(edge_index)
+    assert out[0].tolist() == [[0, 1, 2, 2], [1, 2, 1, 3]]
+    assert out[1].tolist() == [True, False, True, True, True, False]
+
+    torch.manual_seed(6)
+    out = dropout_edge(edge_index, force_undirected=True)
+    assert out[0].tolist() == [[0, 1, 1, 2], [1, 2, 0, 1]]
+    assert out[1].tolist() == [0, 2, 0, 2]