+++ torch_geometric/utils/dropout.py
@@ -3,6 +3,7 @@ from typing import Optional, Tuple
 import torch
 from torch import Tensor
 
+from torch_geometric.deprecation import deprecated
 from torch_geometric.typing import OptTensor
 
 from .num_nodes import maybe_num_nodes
@@ -14,6 +15,7 @@ def filter_adj(row: Tensor, col: Tensor, edge_attr: OptTensor,
     return row[mask], col[mask], None if edge_attr is None else edge_attr[mask]
 
 
+@deprecated("use 'dropout_edge' instead")
 def dropout_adj(
     edge_index: Tensor,
     edge_attr: OptTensor = None,
@@ -25,6 +27,12 @@ def dropout_adj(
     r"""Randomly drops edges from the adjacency matrix
     :obj:`(edge_index, edge_attr)` with probability :obj:`p` using samples from
     a Bernoulli distribution.
+
+    .. warning::
+
+        :class:`~torch_geometric.utils.dropout_adj` is deprecated and will
+        be removed in a future release.
+        Use :class:`torch_geometric.utils.dropout_edge` instead.
 
     Args:
         edge_index (LongTensor): The edge indices.
@@ -92,8 +100,8 @@ def dropout_node(edge_index: Tensor, p: float = 0.5,
     a Bernoulli distribution.
 
     The method returns (1) the retained :obj:`edge_index`, (2) the edge mask
-    indicating which edges were dropped. (3) the node mask indicating
-    which nodes were dropped.
+    indicating which edges were retained. (3) the node mask indicating
+    which nodes were retained.
 
     Args:
         edge_index (LongTensor): The edge indices.
@@ -135,3 +143,68 @@ def dropout_node(edge_index: Tensor, p: float = 0.5,
                                         num_nodes=num_nodes,
                                         return_edge_mask=True)
     return edge_index, edge_mask, node_mask
+
+
+def dropout_edge(edge_index: Tensor, p: float = 0.5,
+                 force_undirected: bool = False,
+                 training: bool = True) -> Tuple[Tensor, Tensor]:
+    r"""Randomly drops edges from the adjacency matrix
+    :obj:`edge_index` with probability :obj:`p` using samples from
+    a Bernoulli distribution.
+
+    The method returns (1) the retained :obj:`edge_index`, (2) the edge mask
+    or index indicating which edges were retained, depending on the argument
+    :obj:`force_undirected`.
+
+    Args:
+        edge_index (LongTensor): The edge indices.
+        p (float, optional): Dropout probability. (default: :obj:`0.5`)
+        force_undirected (bool, optional): If set to :obj:`True`, will either
+            drop or keep both edges of an undirected edge.
+            (default: :obj:`False`)
+        training (bool, optional): If set to :obj:`False`, this operation is a
+            no-op. (default: :obj:`True`)
+
+    :rtype: (:class:`LongTensor`, :class:`BoolTensor` or :class:`LongTensor`)
+
+    Examples:
+
+        >>> edge_index = torch.tensor([[0, 1, 1, 2, 2, 3],
+        ...                            [1, 0, 2, 1, 3, 2]])
+        >>> edge_index, edge_mask = dropout_edge(edge_index)
+        >>> edge_index
+        tensor([[0, 1, 2, 2],
+                [1, 2, 1, 3]])
+        >>> edge_mask # masks indicating which edges are retained
+        tensor([ True, False,  True,  True,  True, False])
+
+        >>> edge_index, edge_id = dropout_edge(edge_index,
+        ...                                      force_undirected=True)
+        >>> edge_index
+        tensor([[0, 1, 2, 1, 2, 3],
+                [1, 2, 3, 0, 1, 2]])
+        >>> edge_id # indices indicating which edges are retained
+        >>> tensor([0, 2, 4, 0, 2, 4])
+    """
+    if p < 0. or p > 1.:
+        raise ValueError(f'Dropout probability has to be between 0 and 1 '
+                         f'(got {p}')
+
+    if not training or p == 0.0:
+        edge_mask = edge_index.new_ones(edge_index.size(1), dtype=torch.bool)
+        return edge_index, edge_mask
+
+    row, col = edge_index
+
+    edge_mask = torch.rand(row.size(0), device=edge_index.device) >= p
+
+    if force_undirected:
+        edge_mask[row > col] = False
+
+    edge_index = edge_index[:, edge_mask]
+
+    if force_undirected:
+        edge_index = torch.cat([edge_index, edge_index.flip(0)], dim=1)
+        edge_mask = edge_mask.nonzero().repeat((2, 1)).squeeze()
+
+    return edge_index, edge_mask